Final Commit: Modularized app by separating routes, forms, and logic for maintainability
