+++ templates/availability_settings.html
@@ -34,7 +34,7 @@
       </div>
 
       <button type="submit" class="btn btn-primary">Save</button>
-      <a href="{{ url_for('tutor_dashboard') }}" class="btn btn-secondary ms-2">Cancel</a>
+      <a href="{{ url_for('tutor_bp.tutor_dashboard') }}" class="btn btn-secondary ms-2">Cancel</a>
     </form>
   </div>
