+++ templates/login.html
@@ -24,7 +24,7 @@
             {% endfor %}
           {% endif %}
         {% endwith %}
-        <form method="post" action="{{ url_for('login') }}">
+        <form method="post" action="{{ url_for('auth_bp.login') }}">
           {{ form.hidden_tag() }}
           <div class="mb-3">
             {{ form.email.label(class="form-label") }}
@@ -44,7 +44,7 @@
         </form>
         <p class="mt-3 text-center">
           Don’t have an account?
-          <a href="{{ url_for('register') }}">Register here</a>.
+          <a href="{{ url_for('auth_bp.register') }}">Register here</a>.
         </p>
       </div>
     </div>
