+++ templates/make_appointment.html
@@ -19,7 +19,7 @@
     {% endif %}
   {% endwith %}
 
-  <form method="POST" action="{{ url_for('make_appointment') }}">
+  <form method="POST" action="{{ url_for('student_bp.make_appointment') }}">
     <div class="mb-3">
       <label for="date" class="form-label">Choose a date</label>
       <input type="date" class="form-control" id="date" name="date" required>
@@ -33,7 +33,7 @@
       <input type="text" class="form-control" id="tutor_filter" name="tutor_filter">
     </div>
     <button type="submit" class="btn btn-primary">Search</button>
-    <a href="{{ url_for('student_dashboard') }}" class="btn btn-secondary">Cancel</a>
+    <a href="{{ url_for('student_bp.student_dashboard') }}" class="btn btn-secondary">Cancel</a>
   </form>
 </div>
