+++ templates/rate_tutor.html
@@ -43,7 +43,7 @@
       </div>
 
       <button type="submit" class="btn btn-primary">Submit Review</button>
-      <a href="{{ url_for('student_dashboard') }}" class="btn btn-secondary ms-2">Cancel</a>
+      <a href="{{ url_for('student_bp.student_dashboard') }}" class="btn btn-secondary ms-2">Cancel</a>
     </form>
   </div>
