+++ templates/search_results.html
@@ -14,7 +14,7 @@
         <li class="list-group-item d-flex justify-content-between align-items-center">
           <div>
             <strong>
-              <a href="{{ url_for('tutor_profile', email=tutor.email) }}">
+              <a href="{{ url_for('student_bp.tutor_profile', email=tutor.email) }}">
                 {{ tutor.name }}
               </a>
             </strong><br>
@@ -26,7 +26,7 @@
               <br>Average Rating: ⭐ {{ tutor.avg_rating }}/5
             {% endif %}
           </div>
-          <form method="GET" action="{{ url_for('select_time') }}">
+          <form method="GET" action="{{ url_for('student_bp.select_time') }}">
             <input type="hidden" name="tutor" value="{{ tutor.email }}">
             <input type="hidden" name="date" value="{{ date }}">
             <input type="hidden" name="subject" value="{{ subject }}">
@@ -39,7 +39,7 @@
     <p>No tutors found matching your criteria.</p>
   {% endif %}
 
-  <a href="{{ url_for('make_appointment') }}" class="btn btn-secondary mt-4">Back</a>
+  <a href="{{ url_for('student_bp.make_appointment') }}" class="btn btn-secondary mt-4">Back</a>
 </div>
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
