+++ templates/test.html
@@ -24,7 +24,7 @@
         {% endfor %}
       </div>
       <button type="submit" class="btn btn-success">Update</button>
-      <a href="{{ url_for('tutor_dashboard') }}" class="btn btn-secondary">Cancel</a>
+      <a href="{{ url_for('tutor_bp.tutor_dashboard') }}" class="btn btn-secondary">Cancel</a>
     </form>
 </div>
 </body>
