+++ templates/tutor_dashboard.html
@@ -34,7 +34,7 @@
         <td>{{ appt.student }}</td>
         <td>{{ appt.subject }}</td>
         <td>
-          <form method="POST" action="{{ url_for('cancel_appointment') }}" class="d-inline">
+          <form method="POST" action="{{ url_for('student_bp.cancel_appointment') }}" class="d-inline">
             <input type="hidden" name="meeting_id" value="{{ appt.id }}">
             <button type="submit" class="btn btn-sm btn-danger"
                     onclick="return confirm('Cancel this appointment?');">
@@ -47,9 +47,9 @@
     </tbody>
   </table>
 
-  <a href="{{ url_for('logout') }}" class="btn btn-secondary">Logout</a>
-  <a href="{{ url_for('account_settings') }}" class="btn btn-outline-primary">Account Settings</a>
-  <a href="{{ url_for('availability_settings') }}" class="btn btn-outline-success">Availability Settings</a>
+  <a href="{{ url_for('auth_bp.logout') }}" class="btn btn-secondary">Logout</a>
+  <a href="{{ url_for('settings_bp.account_settings') }}" class="btn btn-outline-primary">Account Settings</a>
+  <a href="{{ url_for('tutor_bp.availability_settings') }}" class="btn btn-outline-success">Availability Settings</a>
 </div>
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
