+++ templates/tutor_profile.html
@@ -18,7 +18,7 @@
     <p><em>No ratings yet.</em></p>
   {% endif %}
 
-  <a href="{{ url_for('make_appointment') }}" class="btn btn-secondary mt-3">Back to Search</a>
+  <a href="{{ url_for('student_bp.make_appointment') }}" class="btn btn-secondary mt-3">Back to Search</a>
 </div>
 
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
